IMU Support and Kinematic Polish
Largely added support for talking to the IMU and running sensor fusion on the output. Some small streamlining adjustments were maded to the kinematic scripts as well, the largest of which being multithreading to allow the program to do "other useful stuff" while sending positions to the servos (which seems to take a looong time).